Refactor python part of python bindings
